requirements.txt
fastapi>=0.95.0
fastmcp>=0.4.0
uvicorn>=0.21.1
fmpsdk>=0.2.5
python-dotenv>=1.0.0
httpx>=0.24.1
